first set of github reviews
